Storybook: upgrade to v6
core-js added to fix a build reference issue
