--- .storybook/addons.js
@@ -1,23 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-// for docs see https://github.com/echoulen/storybook-addon-styled-component-theme
-import 'storybook-addon-styled-component-theme/dist/register'
-
-// for docs see https://www.npmjs.com/package/@storybook/addon-backgrounds
-import '@storybook/addon-backgrounds/register'
-
-// for docs see https://github.com/storybooks/storybook/tree/master/addons/actions
-import '@storybook/addon-actions/register'
-
-// for docs see https://github.com/storybooks/storybook/tree/master/addons/knobs
-import '@storybook/addon-knobs/register'
-
-// for docs see https://github.com/storybooks/storybook/tree/master/addons/options
-import '@storybook/addon-options/register'
-
-// for docs see https://github.com/dump247/storybook-state
-// this addon is surfacing in every panel ATM. this is a bug. track it here:
-// https://github.com/dump247/storybook-state/pull/12
-// import '@dump247/storybook-state/register'
