+++ .storybook/main.js
@@ -1,0 +1,7 @@
+module.exports = {
+  stories: ['../components/**/stories/*.tsx'],
+  addons: [
+    '@storybook/addon-knobs',
+    '@storybook/addon-essentials'
+  ]
+}
