+++ components/brave_new_tab_ui/stories/default/data/brandedWallpaper.ts
@@ -1,9 +1,9 @@
 import wallpaperImageUrl from '../../../../img/newtab/dummy-branded-wallpaper/background-1.jpg'
 import brandingImageUrl from '../../../../img/newtab/dummy-branded-wallpaper/logo.png'
 
-export default {
-  wallpaperImageUrl,
+const dummyWallpaper: NewTab.BrandedWallpaper = {
   isSponsored: true,
+  wallpaperImageUrl,
   creativeInstanceId: '12345abcde',
   wallpaperId: 'abcde12345',
   logo: {
@@ -13,3 +13,5 @@ export default {
     destinationUrl: 'https://brave.com'
   }
 }
+
+export default dummyWallpaper
