+++ components/brave_new_tab_ui/stories/privateNTP/privateWindow.tsx
