+++ components/brave_welcome_ui/stories/story.tsx
@@ -3,32 +3,24 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import * as React from 'react'
-import { storiesOf } from '@storybook/react'
 import welcomeDarkTheme from '../theme/welcome-dark'
 import welcomeLightTheme from '../theme/welcome-light'
-import { withThemesProvider } from 'storybook-addon-styled-component-theme'
-import { withKnobs, boolean } from '@storybook/addon-knobs'
+import { boolean } from '@storybook/addon-knobs'
 
 // Components
 import WelcomePage from './page/index'
 
-// Themes
-const themes = [welcomeLightTheme, welcomeDarkTheme]
-
-const fullPageStoryStyles: object = {
-  width: '100%',
-  height: '100%'
+export default {
+  title: 'Welcome',
+  args: {
+    darkTheme: welcomeDarkTheme,
+    lightTheme: welcomeLightTheme
+  },
+  parameters: {
+    layout: 'fullscreen'
+  }
 }
 
-export const FullPageStory = (storyFn: any) =>
-  <div style={fullPageStoryStyles}>{storyFn()}</div>
-
-storiesOf('Welcome', module)
-  .addDecorator(withThemesProvider(themes))
-  .addDecorator(withKnobs)
-  .addDecorator(FullPageStory)
-  .add('Page', () => {
-    return (
-      <WelcomePage isDefaultSearchGoogle={boolean('Is default search google?', true)}/>
-    )
-  })
+export const Page = () => (
+  <WelcomePage isDefaultSearchGoogle={boolean('Is default search google?', true)}/>
+)
