+++ components/webcompat_reporter/ui/stories/story.tsx
@@ -4,46 +4,54 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import * as React from 'react'
-import { storiesOf } from '@storybook/react'
-import Theme from 'brave-ui/theme/brave-default'
-import DarkTheme from 'brave-ui/theme/brave-dark'
-import { withThemesProvider } from 'storybook-addon-styled-component-theme'
-// @ts-ignore
-import { boolean, number } from '@storybook/addon-knobs'
-
-// Components
 import WebcompatReportModal from './modal/index'
 
-// Themes
-const themes = [Theme, DarkTheme]
+export default {
+  title: 'Webcompat Reporter'
+}
 
-storiesOf('Webcompat Reporter', module)
-  .addDecorator(withThemesProvider(themes))
-  .add('Before submit', () => {
-    const fakeOnSubmit = () => { /* noop */ }
-    const fakeOnClose = () => { /* noop */ }
-    return (
-      <div style={{ width: '375px', margin: '0 auto' }}>
-        <WebcompatReportModal
-          siteUrl={'https://www.buzzfeed.com/'}
-          submitted={false}
-          onSubmitReport={fakeOnSubmit}
-          onClose={fakeOnClose}
-        />
-      </div>
-    )
-  })
-  .add('Submit confirmation', () => {
-    const fakeOnSubmit = () => { /* noop */ }
-    const fakeOnClose = () => { /* noop */ }
-    return (
-      <div style={{ width: '375px', margin: '0 auto' }}>
-        <WebcompatReportModal
-          siteUrl={'https://www.buzzfeed.com/'}
-          submitted={true}
-          onSubmitReport={fakeOnSubmit}
-          onClose={fakeOnClose}
-        />
-      </div>
-    )
-  })
+export const BeforeSubmit = () => {
+  const fakeOnSubmit = () => {
+    /* noop */
+  }
+  const fakeOnClose = () => {
+    /* noop */
+  }
+  return (
+    <div style={{ width: '375px', margin: '0 auto' }}>
+      <WebcompatReportModal
+        siteUrl={'https://www.buzzfeed.com/'}
+        submitted={false}
+        onSubmitReport={fakeOnSubmit}
+        onClose={fakeOnClose}
+      />
+    </div>
+  )
+}
+
+BeforeSubmit.story = {
+  name: 'Before submit'
+}
+
+export const SubmitConfirmation = () => {
+  const fakeOnSubmit = () => {
+    /* noop */
+  }
+  const fakeOnClose = () => {
+    /* noop */
+  }
+  return (
+    <div style={{ width: '375px', margin: '0 auto' }}>
+      <WebcompatReportModal
+        siteUrl={'https://www.buzzfeed.com/'}
+        submitted={true}
+        onSubmitReport={fakeOnSubmit}
+        onClose={fakeOnClose}
+      />
+    </div>
+  )
+}
+
+SubmitConfirmation.story = {
+  name: 'Submit confirmation'
+}
