+++ package.json
@@ -240,15 +240,12 @@
   },
   "homepage": "https:/github.com/brave/brave-core#readme",
   "devDependencies": {
-    "@babel/core": "7.12.16",
+    "@babel/core": "^7.13.1",
     "@dump247/storybook-state": "1.6.1",
-    "@storybook/addon-actions": "5.3.21",
-    "@storybook/addon-backgrounds": "5.3.21",
-    "@storybook/addon-centered": "5.3.21",
-    "@storybook/addon-knobs": "5.3.21",
-    "@storybook/addon-options": "5.3.21",
-    "@storybook/addons": "5.3.21",
-    "@storybook/react": "5.3.21",
+    "@storybook/addon-essentials": "^6.1.21",
+    "@storybook/addon-knobs": "^6.1.11",
+    "@storybook/addons": "^6.1.11",
+    "@storybook/react": "^6.1.11",
     "@types/array-move": "2.0.0",
     "@types/bluebird": "3.5.33",
     "@types/chrome": "0.0.130",
@@ -291,7 +288,6 @@
     "react-redux": "7.2.2",
     "redux": "4.0.5",
     "sinon": "7.5.0",
-    "storybook-addon-styled-component-theme": "1.3.0",
     "style-loader": "0.23.1",
     "styled-components": "3.4.10",
     "ts-jest": "25.5.1",
@@ -314,6 +310,7 @@
     "array-move": "^2.2.1",
     "bignumber.js": "^7.2.1",
     "bluebird": "^3.5.1",
+    "core-js": "^3.9.1",
     "date-fns": "^2.15.0",
     "jszip": "^3.2.2",
     "parse-domain": "^3.0.2",
